move metadata_comparison under operators dir; move related tests into metadata_comparison_test
