+++ test/models/concerns/advanced_search/operators/comparison_operators_test.rb
@@ -8,7 +8,7 @@ module AdvancedSearch
       # Test class that includes the ComparisonOperators module
       class TestClass
         include AdvancedSearch::Operators::ComparisonOperators
-        include AdvancedSearch::MetadataComparison
+        include AdvancedSearch::Operators::MetadataComparison
       end
 
       def setup
@@ -96,50 +96,6 @@ module AdvancedSearch
         assert_includes sql, 'CAST'
         assert_includes sql, 'DOUBLE PRECISION'
         assert_includes sql, '>='
-      end
-
-      # condition_date_comparison tests
-      test 'condition_date_comparison includes date regex validation' do
-        result = @test_instance.send(:metadata_condition_date_comparison, @scope, @node, '2024-01-01', :lteq)
-        sql = result.to_sql
-        # Regex matches YYYY, YYYY-MM, or YYYY-MM-DD formats
-        assert_includes sql, '~'
-        assert_includes sql, 'TO_DATE'
-        assert_includes sql, 'YYYY-MM-DD'
-      end
-
-      test 'condition_date_comparison returns none for invalid date format' do
-        result = @test_instance.send(:metadata_condition_date_comparison, @scope, @node, '2024-99-40', :lteq)
-        assert result.none?
-      end
-
-      test 'condition_date_comparison works with gteq' do
-        result = @test_instance.send(:metadata_condition_date_comparison, @scope, @node, '2024-06-15', :gteq)
-        sql = result.to_sql
-        assert_includes sql, '>='
-        assert_includes sql, 'TO_DATE'
-      end
-
-      # condition_numeric_comparison tests
-      test 'condition_numeric_comparison includes numeric regex validation' do
-        result = @test_instance.send(:metadata_condition_numeric_comparison, @scope, @node, '123', :lteq)
-        sql = result.to_sql
-        # Regex matches integers and decimals (including negatives)
-        assert_includes sql, '~'
-        assert_includes sql, 'CAST'
-      end
-
-      test 'condition_numeric_comparison returns none for invalid numeric format' do
-        result = @test_instance.send(:metadata_condition_numeric_comparison, @scope, @node, '12a3', :lteq)
-        assert result.none?
-      end
-
-      test 'condition_numeric_comparison works with gteq' do
-        result = @test_instance.send(:metadata_condition_numeric_comparison, @scope, @node, '456.78', :gteq)
-        sql = result.to_sql
-        assert_includes sql, '>='
-        assert_includes sql, 'CAST'
-        assert_includes sql, 'DOUBLE PRECISION'
       end
     end
   end
